+++ src/SqlArtisan/Internal/SqlPart/Clause/ForUpdate/LockBehaviorBase.cs
@@ -2,4 +2,7 @@ namespace SqlArtisan.Internal;
 
 public abstract class LockBehaviorBase : SqlPart
 {
+    private protected LockBehaviorBase()
+    {
+    }
 }